networking: Refactor the module structure.
- gather custom protocols and request handlers in the “protocol” module
