+++ crates/subspace-networking/src/protocols/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod connected_peers;
+pub mod peer_info;
+pub mod requests;
+pub(crate) mod reserved_peers;
